+++ libs/rxjs/number/src/index.ts
@@ -8,6 +8,7 @@
  * @ignore
  */
 /* istanbul ignore file */
+export { add } from './lib/add';
 export { filterInRange } from './lib/filter-in-range';
 export { filterIsFinite } from './lib/filter-is-finite';
 export { filterIsFloat } from './lib/filter-is-float';
@@ -29,7 +30,7 @@ export { outOfRange } from './lib/out-of-range';
 export { parseFloat } from './lib/parse-float';
 export { parseHex } from './lib/parse-hex';
 export { parseInt } from './lib/parse-int';
-export { raiseBy } from './lib/raise-by';
+export { pow } from './lib/pow';
 export { toExponential } from './lib/to-exponential';
 export { toFixed } from './lib/to-fixed';
 export { toHex } from './lib/to-hex';